requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-genai
python-multipart
slowapi
fpdf2
pandas
openpyxl
httpx
websockets
jinja2
psycopg2-binary
schedule
python-dotenv
pydantic-settings
playwright
